fix: Better way to deal with data urls.
